Remove code using a deprecated method (#120)
* Enable deprecation lint check

* Remove code using a deprecated method

The `getRegisters()` method of `SDT` has been marked deprecated for a
long time, but has been kept in the code base as it is used in various
tests.  This commit makes the method private and removes (unused) code
in other classes that was calling this method.

This change shuts off two deprecation warnings from the linter.